@@ -11,18 +11,22 @@
     "array-move": "^2.1.0",
     "babel-preset-es2018": "^1.0.0",
     "bootstrap": "^4.1.3",
+    "capitalize": "^2.0.0",
     "classnames": "^2.2.6",
     "connected-react-router": "^6.3.2",
     "css-modules": "^0.3.0",
     "dot-prop-immutable": "^1.5.0",
     "flow-bin": "^0.95.1",
+    "html2canvas": "^1.0.0-alpha.12",
+    "jspdf": "^1.5.3",
     "moment": "^2.24.0",
     "node-sass": "^4.11.0",
     "query-string": "^6.7.0",
-    "react": "^16.6.0",
+    "react": "^16.8.6",
     "react-content-loader": "^4.2.1",
     "react-contenteditable": "^3.2.6",
     "react-cookie": "^4.0.0",
+    "react-copy-to-clipboard": "^5.0.1",
     "react-csv": "^1.1.1",
     "react-dom": "^16.6.0",
     "react-hot-loader": "^4.8.2",
