@@ -1,11 +1,9 @@
 import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
-import { getStatementsByObject } from '../network';
+import { getStatementsByObject, getStatementsBySubject } from '../network';
 import { Container } from 'reactstrap';
-import { reverse } from 'named-urls';
-import ROUTES from '../constants/routes.js';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faSpinner } from '@fortawesome/free-solid-svg-icons';
+import PaperCard from './../components/PaperCard/PaperCard'
 /*
 This page is only for debugging. It is requesting ALL statements and resources in 
 order to filter out non-paper resources (very inefficient)
@@ -16,14 +14,46 @@ export default class Resources extends Component {
         results: null,
     };
 
-    async componentWillMount() {
-        let statements = await getStatementsByObject({
+    componentDidMount() {
+        getStatementsByObject({
             id: process.env.REACT_APP_RESOURCE_TYPES_PAPER,
             order: 'desc',
-        });
-
-        this.setState({
-            statements
+        }).then((papers) => {
+            // Fetch the data of each paper
+            var papers_data = papers.map((paper) => {
+                return getStatementsBySubject(paper.subject.id).then((paperStatements) => {
+                    // publication year
+                    let publicationYear = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_PUBLICATION_YEAR);
+                    if (publicationYear.length > 0) {
+                        publicationYear = publicationYear[0].object.label
+                    }
+                    // publication month
+                    let publicationMonth = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_PUBLICATION_MONTH);
+                    if (publicationMonth.length > 0) {
+                        publicationMonth = publicationMonth[0].object.label
+                    }
+                    // authors
+                    let authors = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_AUTHOR);
+                    let authorNamesArray = [];
+                    if (authors.length > 0) {
+                        for (let author of authors) {
+                            let authorName = author.object.label;
+                            authorNamesArray.push(authorName);
+                        }
+                    }
+                    paper.data = {
+                        publicationYear,
+                        publicationMonth,
+                        authorNames: authorNamesArray.reverse(),
+                    }
+                    return paper;
+                })
+            });
+            return Promise.all(papers_data).then((statements) => {
+                this.setState({
+                    statements
+                });
+            })
         });
     }
 
@@ -33,19 +63,20 @@ export default class Resources extends Component {
                 <Container className="p-0">
                     <h1 className="h4 mt-4 mb-4">View all papers</h1>
                 </Container>
-                <Container className="box pt-4 pb-4 pl-5 pr-5">
+                <Container className={'p-0'}>
                     {this.state.statements && this.state.statements.length > 0 ?
-                        <ul className="list-group list-group-flush">
+                        <div>
                             {this.state.statements.map(
-                                resource => (
-                                    <li className="list-group-item list-group-item-action" key={resource.id}>
-                                        <Link to={reverse(ROUTES.VIEW_PAPER, { resourceId: resource.subject.id })}>
-                                            {`${resource.subject.id}: ${resource.subject.label}`}
-                                        </Link>
-                                    </li>
-                                )
+                                (resource) => {
+                                    return (
+                                        <PaperCard
+                                            paper={{ id: resource.subject.id, title: resource.subject.label, ...resource.data }}
+                                            key={`pc${resource.id}`}
+                                        />
+                                    )
+                                }
                             )}
-                        </ul>
+                        </div>
                         : !this.state.statements && <div className="text-center mt-4 mb-4"><Icon icon={faSpinner} spin /> Loading</div>}
                 </Container>
             </>
