@@ -35,7 +35,7 @@ class ResearchProblemInput extends Component {
             )
             // Set them to the list of research problems and add the created options
             this.setState({
-                researchProblems: [...researchProblems, this.props.value.filter(({ id }) => !researchProblems.map(({ id }) => id).includes(id))],
+                researchProblems: [...researchProblems, ...this.props.value.filter(({ id }) => !researchProblems.map(({ id }) => id).includes(id))],
                 loading: false
             })
         })
@@ -67,9 +67,13 @@ class ResearchProblemInput extends Component {
                 paddingLeft: 0,
                 paddingRight: 0,
             }),
-            multiValue: (provided, state) => ({
+            multiValue: (provided) => ({
                 ...provided
             }),
+            menu: (provided) => ({
+                ...provided,
+                zIndex:10
+            })
         }
 
         const Menu = props => {
