@@ -6,6 +6,8 @@ import styled from 'styled-components';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faSort } from '@fortawesome/free-solid-svg-icons';
 import { SortableContainer, SortableElement, sortableHandle } from 'react-sortable-hoc';
+import capitalize from 'capitalize';
+import Tooltip from '../Utils/Tooltip';
 
 const DragHandle = styled.span`
     cursor:move;
@@ -37,13 +39,15 @@ class SelectProperties extends Component {
             {property.active ? <this.SortableHandle /> : <DragHandlePlaceholder />}
             <CustomInput
                 type="checkbox"
-                id={`checkbox-${property.label}`}
-                label={property.label}
+                id={`checkbox-${property.id}`}
+                label={capitalize(property.label)}
                 className="flex-grow-1"
                 onChange={() => this.props.toggleProperty(property.id)}
                 checked={property.active}
             />
-            <Badge color="lightblue">3</Badge>
+            <Tooltip message="Amount of contributions" hideDefaultIcon>
+                <Badge color="lightblue">{property.contributionAmount}</Badge>
+            </Tooltip>
         </ListGroupItemStyled>
     ));
     
