@@ -1,14 +1,15 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { selectContribution } from '../../actions/viewPaper';
-
+import { Link } from 'react-router-dom';
+import { reverse } from 'named-urls';
+import PropTypes from 'prop-types';
 import styled from 'styled-components';
+import ROUTES from '../../constants/routes.js';
 
 const CardContainer = styled.div`
     display: flex;
 `;
 
-const Card = styled.div`
+const Card = styled(Link)`
     border:2px;
     width:calc(33% - 5px);
     flex-grow:0;
@@ -19,9 +20,20 @@ const Card = styled.div`
     cursor:pointer;
     font-size:95%;
     padding:10px 5px;
+    color:${props => props.theme.bodyColor};
+    text-decoration:none;
 
     &:last-child {
         margin-right:0;
+    }
+    
+    &:hover{
+        color:${props => props.theme.orkgPrimaryColor};
+        text-decoration:none;
+    }  
+
+    &, &:active, &:focus{
+        outline: none;
     }
 `;
 
@@ -29,7 +41,6 @@ const Similarity = styled.span`
     width:35px;
     border-right:2px solid;
     border-right-color:inherit;
-    height:100%;
     display:block;
     float:left;
     text-align:center;
@@ -40,37 +51,35 @@ const Similarity = styled.span`
 
 
 class SimilarContributions extends Component {
+
     render() {
 
         return (
-            <CardContainer>
-                <Card className="justify-content-center" role="button">
-                    <Similarity>80<br />%</Similarity>
-                    Wiles’s proof of Fermat’s last theorem
-                </Card>
-                <Card className="justify-content-center" role="button">
-                    <Similarity>54<br />%</Similarity>
-                    Gruber’s design of ontologies
-                </Card>
-                <Card className="justify-content-center" role="button">
-                    <Similarity>14<br />%</Similarity>
-                    Design criteria for ontologies
-                </Card>
-            </CardContainer>
+            <>
+                {this.props.similaireContributions.length > 0 && (
+                    <CardContainer>
+                        {this.props.similaireContributions.map((contribution, index) => {
+                            return (
+                                <Card
+                                    key={`s${contribution.contributionId}`}
+                                    to={reverse(ROUTES.VIEW_PAPER_CONTRIBUTION, { resourceId: contribution.paperId, contributionId: contribution.contributionId })}
+                                    className="justify-content-center"
+                                    role="button"
+                                >
+                                    <Similarity>{contribution.similarityPercentage}<br />%</Similarity>
+                                    {contribution.title}
+                                </Card>
+                            )
+                        })}
+                    </CardContainer>
+                )}
+            </>
         );
     }
 }
 
-const mapStateToProps = state => ({
-    viewPaper: state.viewPaper,
-    addPaper: state.addPaper,
-});
+SimilarContributions.propTypes = {
+    similaireContributions: PropTypes.array.isRequired,
+};
 
-const mapDispatchToProps = dispatch => ({
-    selectContribution: (data) => dispatch(selectContribution(data)),
-});
-
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(SimilarContributions);
+export default (SimilarContributions);
